# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for GoApi.",
        "title": "GoApi Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/GetAddressBookByUserName": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get by username address books",
                "produces": [
                    "application/json"
                ],
                "summary": "Get by username address books",
                "operationId": "get-address-book-by-username",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AddressBook"
                            }
                        }
                    }
                }
            }
        },
        "/getAddress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all address books",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all address books",
                "operationId": "get-address-book-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AddressBook"
                            }
                        }
                    }
                }
            }
        },
        "/getToken": {
            "get": {
                "description": "generate a JWT token for a given username",
                "produces": [
                    "application/json"
                ],
                "summary": "Generate JWT token",
                "operationId": "generate-token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddressBook": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}